feat(networking): handle TLS parameter in URL construction
This commit adds support for the 'tls' parameter in the URL construction logic of the networking module. If the 'tls' parameter is set to true, the port used will be 443; otherwise, it will default to port 80. This change allows for more flexible and configurable URL generation based on the presence and value of the 'tls' parameter.

The previous implementation only considered the 'port' parameter, which limited the ability to specify whether TLS should be used or not.